requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-mail==0.9.1
flask-cors==4.0.0
werkzeug==3.0.0
Pillow==10.0.0
python-dotenv==1.0.0
pillow-heif==0.7.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
flask-limiter==3.5.0